# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/webhooks/paymongo": {
      "post": {
        "operationId": "PaymongoWebhookController_handle",
        "parameters": [
          {
            "name": "paymongo-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["webhooks"]
      }
    },
...
  },
  "info": {
    "title": "ClinIQ API",
    "description": "Multi-tenant clinic management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4005"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}